feat: [PartsCollection 페이지]  비로그인 예외처리
