+++ packages/service/src/apis/partsEvent/apiGetMyParts.ts
@@ -1,9 +1,10 @@
+import { customFetch } from "@service/common/utils/customFetch";
 import { IMyParts } from "./type";
 
 export const apiGetMyParts: (token: string) => Promise<IMyParts[]> = (
   token,
 ) => {
-  return fetch(`${import.meta.env.VITE_BACK_BASE_URL}/event/parts`, {
+  return customFetch(`${import.meta.env.VITE_BACK_BASE_URL}/event/parts`, {
     headers: {
       Authorization: `Bearer ${token}`,
     },
